backend/auth/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8
asyncpg>=0.28.0
sqlalchemy>=2.0.0
alembic>=1.12.0
httpx>=0.25.0
authlib>=1.2.0
itsdangerous>=2.1.0
cryptography>=41.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
pyotp>=2.9.0
qrcode>=7.4.0
pillow>=10.0.0
